EmployeePortal/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
idna==3.4
numpy==1.25.2
opencv-python==4.8.0.74
Pillow==10.0.0
platformdirs==3.9.1
psycopg2==2.9.6
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.2
